+++ src/format_null.c
@@ -45,5 +45,5 @@ static FILE *open_for_output(char *filename, proc_info *pinfo) {
 }
 
 static void create_output_filename(char *outfilename) {
-  strcpy(outfilename, NULLDEVICE);
+  st_strlcpy(outfilename, NULLDEVICE, FILENAME_SIZE);
 }